requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
# Editable Git install with no remote (atm_interface==0.1)
-e a:\ramspheld\brainwave_matrix_intern\atm_interface
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
Flask>=2.0.1
flask-swagger-ui>=3.36.0
frozenlist==1.6.2
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pydantic>=2.0.0
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest>=6.2.5
pytest-asyncio==1.0.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
PyYAML>=5.4.1
realtime==2.4.3
redis==6.2.0
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase>=0.0.2
supafunc==0.9.4
typing-inspection==0.4.1
typing_extensions==4.14.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
gunicorn>=20.1.0
bcrypt>=4.0.1
python-jose>=3.3.0
